Use sane min-max ranges for spinboxes
Qt 5.1 determines the minimum size hint for a spinbox from the space taken by
the digits of the min and max values, cf.
https://qt.gitorious.org/qt/qtbase/commit/a317ee0a6fa76d1166f6da8593d39eaf7afce83c
The values returned by std::numeric_limits for double etc. take too much space
and cannot be used.

This commit adds a minimal infrastructure similar to std::numeric_limits for
representing type properties which are relevant for input widgets.