Don't use binary shift on signed integers
Replace both left-shift and right-shift of signed integers with
multiplication and division, respectively.

Binary shift of signed integers already was implementation-defined
behavior in C++. It is even undefined behavior when a signed integer
arithmetic operation (such as binary left-shift) overflows, cf.
http://en.cppreference.com/w/cpp/language/operator_arithmetic

C++11 constexpr must not contain UB. Binary left-shift of signed
integers in a constexpr is a compile-time error at least in GCC 5.1.