Georeferencing: Use +no_defs for all Proj.4 specifications
The defaults file proj_def.dat doesn't add any value for most
specifications. It may even lead to surprising errors. In proj-4.9.1,
the file was missing. This resulted in an error when using UTM in Mapper,
uncovered by GeoreferencingTest. The test passes again with this change.
At least Ubuntu 15.10 was affected by this error