TemplateMap: Calculate transformations for new maps
When adding a new template map, a transformation must be calculated
from the georeferencings of the active map and the template map. This
transformation can be determined from a small number of pass points
which are mapped by means of Georeferencing.
Existing template maps are assumed to have a valid transformation.
Fixes #784.