Fix QTransform usage
PassPointList::estimateNonIsometricSimilarityTransform and
qTransformToTemplateTransform used m13 and m23 of QTransform for storing
and reading the translation parameters. But QTransform is documented to
store translation in m31 and m32. There was no visible bug yet because
the functions were used together. But an attempt to directly use the
QTransform created by the first function run into trouble.