TranslationUtil: Major rewrite
For translation, no longer rely on being able to create instances of
QLocale for all QLocale::Language values. This makes it possible to use
the Esperanto language (fixes #792).
Get rid of QMap for the small set of languages.
Save the language code instead of the QLocale::Language numerical value
in the settings, and migrate existing settings.
Add some tests to catch behavioral changes of QLocale.