FileFormatTest: Fix detection of high coordinate import bugs
The test failed to detect the bugs it was made for: Objects with
out-of-bounds coordinates were deleted for being irregular before
the actual tests.
Now it verifies the presence of the map object which is expected to
get out-of-bounds treatment.
Another test file verifies out-of-bounds-treatment for the dense text
format used in omap files. (The test file is identical to the xmap
test file except for the format of the map object's coordinates.)