ObjectQuery: Switch to value semantics
Moving the contents is cheap. Copying the contents is cheap for tag
queries and possible for sub-queries. So we add the constructors and
operators which allow this class to be used with value semantics.
For memory allocations, we still use make_unique instead of new.

In addition, connecting sub-queries will now make the resulting query
invalid if any of the sub-queries is invalid.