MapEditorToolBase: Decouple edited objects from selection
Unless carefully deactivating all actions which might modify the selection
while editing is in progress, relying on a 1:1 mapping of undo duplicates
and selected objects is dangerous. This change makes the tool explicitly
track each active (i.e. edited) object together with the duplicate (backup).
The extra memory is negligible in comparison to the duplicates' size.
In contrast locality of some operations is improved by using a single list.

This change also adds a method which compares the edited objects with the
duplicates, making it possible to avoid the creation of a backup step if
there is no change.

Note that the subclasses carry their own dependencies on the selection
and need to be decoupled separately.