OgrTemplate: Handle offsets and TemplateTrack data
The CRS specification needs to be kept for compatibility with the
TemplateTrack implementation.

The offset for out-of-bounds coordinates which is applied during import
must not affect the saved transformation data because on next loading of
the template, the same offset will be calculated and applied again.
In addition, opening the template with the TemplateTrack class (older
version of Mapper, or OgrTemplate disabled) will not handle offsets.