Add CRT file support for DXF import.
The GDAL/OGR based import makes the DXF layer available in object tags.
This change looks for a corresponding cross reference table file (aka
CRT file) when importing a DXF file. The CRT file must be in the same
directory and have the same base name as the DXF file.

When a CRT file is found, the symbol set replacement dialog is used for
preflight check and manipulation of the symbol replacements.
If the dialog is canceled, it is still optionally possible to import
without symbol manipulations.

Resolves #443. Test data in #261.