Require C++14.
C++14 adds the std::make_unique which complements unique_ptr and auto.

Requiring C++14 excludes GCC 4.8 which implements C++11 but also a
misfeature which results in crashes (initializing members which are
references by creating temporaries when using uniform initialization),
cf. http://stackoverflow.com/questions/25561387/spurious-warning-about-binding-temporary-to-reference-member-in-constructor

For simplicity, C++14 is required by means of CMake >= 3.1. On systems
with older CMake, the user needs to explicitly configure the project
with the required CMAKE_CXX_FLAGS.