Integrate androiddeployqt into CMake package target
There is no suitable CPack generator, but we can hook into
CPackConfig.cmake to present an emtpy list of generators
to CPack, and trigger androiddeployqt at the same time.

The actual configuration and run of androiddeployqt is deferred
to a qmake project which receives extra configuration from
custom_install.cmake.

Mapper must be built as a shared library for Android, and
installation destinations need to be adjusted.
The installation of Qt plugins is left to androiddeployqt.

Tested with OpenOrienteering superbuild.