OCAD8FileExport: Don't access text objects twice
The export may use temporary clones instead of the original objects.
Thus it can be illegal to access pointers to (cloned) text objects
which where processed before the current one. However, for identifying
the exported symbol matching the current horizontal alignment, this
alignment can be used directly instead of the text object pointer.