OgrFileImport: Review SRS initalization and import
There is no point in early initialization of map_srs:
Either it is to be initialized from the data, or we must catch changes
to the map georeferencing made after the construction of the importer.

In addition, the GDAL OSR tutorial strongly recommends a particular
order of steps for defining a projected coordinate system.