Prevent early template loading during import
File formats which use the public API for add templates to maps could
trigger immediate template loading when setting the visiblity in
MapView. But during import, templates are to be loaded in a separate
step, for better reporting of issues. That is why early loading must
be prevented.
Uncovered by an assertion when loading OCD files.