FillTool: Ignore area fill patterns
The fill tool operates on the paths that define lines and areas.
The regions which are colored by area fill patterns have no useful
relation to the outline of the respective area. Assuming that these
patterns produce sparse coloring, it is more reasonable to ignore
them then to treat them as opaque.
Fixes application of the fill tool in areas covered e.g. by a north
line pattern.