MainWindow: Fix loadWindowSettings()
Set windows state instead of calling showMaximized(). This avoids of the
side effect of actually showing the window which is not desired here.
In turn, this avoids triggering a warning from the minimal platform plugin
in automated tests. The warning would cause an abort when QT_FATAL_WARNINGS
is set.