FileFormat: Review return type of understands()
The bool return value was insufficient to indicate situation where it
is not possible to determine whether file data is supported or not.
This change introduces a three-valued enum, elaborates the actual
implementations, and adds a unit test.

Map::loadFrom() no longer attempts to use the next format after
failure with a file format which indicated to fully support the file.
This prevents meaningful error messages being overwritten by formats
such as GDAL which are unable to indicate anything else then unknown
support.