Symbol classes: Make duplicate() return unique_ptr
Class Symbol is polymorphic. Due to the incompatiblity of unique_ptr
with covariant return types, the regular member function duplicate
continues to return (covariant) raw pointers, but becomes protected.
A public static template member function and a free template function
provide a public API which returns unique_ptr.
The factory methods makeSymbolForType(...) and load(...) now return
unique_ptr, too.

The scope of this commit is intentionally limited to class Symbol
and direct dependencies. The full benefit of this change unfolds
with changes to other classes which consume and own Symbol instances.