LineSymbol: Refactor creation of start/end symbols
Mapper used to create a start symbol at the start of the first part
and an end symbol at the end of the last part, respectively, as many
times as there were parts in a path object - i.e. possibly creating
multiple instance in the same location. This change moves the creation
of start/end symbols to a separate function which is called only once
for the whole list of parts of a path object.

This change also moves the (check for) start/end symbols out of the
code path of creating border lines.

Note that Mapper doesn't allow to create lines with multiple parts
directly. Multiple parts for a line symbol may occur only in the
context of combined symbols.