SymbolTest: Start to verify render output
The rendering implementations still need a lot of work, for cleanup
and for further development. Comparing images for purposeful test
files with images with images taken at an earlier point of development
lets us catch unintented changes in rendering. Intented changes may
be documented by updates to the saved images.