TextObject: Move box size out of coords
If a text object is given a fixed size, this size cannot be kept in
the coords structure since they cannot be transformed in the same way
as regular coordinates.
For now, we move the size to an explicit member variable, but use
coords[1] during loading and saving our XML format. In loading, we
prevent the size coordinate from going through bounds handling.
Fixes #1069.

This change lets the legacy native format as is.