Object: Refactor transform()
Transform also path object pattern origin. Instead of adding an if
statement in the common implementation, make this pure virtual in
Object, and add individual implementations in PathObject, PointObject
and TextObject.
Assumed to fixe the root cause of #1061.
This function is called only during map part import and OGR import
at the moment.