translations: Preserve Hungarian plural forms
Weblate provides singular (n=1) and plural translations for Hungarian.
Qt does not support this distinction and would drop the second (i.e.
plural) translation, which is just wrong. The changes in this commit:
- ensure that the plural is used when building,
- maintain the singular translation when updating,
- add a comment to inform translators.
(For verifying the usage of plural, copy some object to the clipboard
and check the status bar.)