PaintOnTemplateTool: Fix button state after template change
Upon construction and (delayed) destruction, tools update their
associated checkable action. So when switching to a new instance
of the same tool, the action will end up in unchecked state.
For switching templates, this is solved in this change by no longer
creating a new tool, but update the existing one instead.