LineSymbol: Implement alternative mid symbol placement
So far, mid symbol were placed in the middle of the dash which
actually meant the middle of the dash group. This change turns the
mid symbol into a more flexible "style symbol" (without changing
the name) which alternatively can be placed in the middle of every
dash in dash groups, or at the gap between dash groups.

This change adds a new attribute to the file format.