OCD: Rework OCD grid/zone lookup
The code is currently designed to handle zones with EPSG codes assigned.  Some
of the zones have multiple possible representations in Mapper realm.  We are
choosing the first match, which is also the compatibility option.

UTM south zones are encoded as negative numbers. Signum is held in zone_id
number so that grid_id values correspond to its enum.