Exporter: Use const Map and View parameters
Move the stream, map and view members from ImportExport to Importer and
Exporter, and make map and view const members in the Exporter.
Accordingly, let the exporter constructor take const map and view
parameters. This reduce the ImportExport base clase to handling logging
and options.
Resolve the remaining spots in exporter implementations.