Importer: Use path instead of device in constructor
This change makes it possible
a) to implement importers which do not use QIODevice at all, and
b) to encapsulate the regular QFile handling in the Importer class
   instead of the Map class, and
c) to have the path available when needed, e.g. for determining
   relative paths of template files.

Similar to the changes done in Exporter before, doImport() will now
catch exceptions and return false on error instead. Error handling
around calls to doImport() is adjusted accordingly.