Remove the setting for choosing the OCD 6-8 importer
This feature is obsolete: The legacy importer may be chosen in
as a filter in the File > Open... dialog.

It is no longer possible to use the old importer with files opened in
another way (recent file list, program argument, template). However,
this is a minor issue. The old importer is unmaintained since a while.
It shall be removed soon.