Review use of PathObject::getCoordinate
The non-const implementation of this function sets the output_dirty flag.
This triggers the expensive recalculation of detailed path data. This is
not needed in many contexts, so we need to make sure to call the const
implementation by appropriate cast of the PathObject.
Both const and non-const implementation return a reference. Where
possible, we can capture this as a reference instead of as a copy.