LineSymbol: Remove pointed caps support from dashes
The original implementation of "pointed line caps" was broken for dashed
lines, and it created different dash layouts than common line caps.
While it would be possible to create a working, regular implementation,
there seem to be no actual use of such a feature. Instead, this commit
turns pointed ends into a feature of the *path* start and end. Dashed
lines with pointed start and end of the path will use flat line caps.

With this change, pure offsets at path start and end (#1125) can be
implemented as a minimal variation, by simply not drawing the pointed
ends but otherwise using the same layout.

This commit does not modify the line symbol settings UI.