LineSymbol: Add support for offset at start/end
This change adds support for line symbols which are drawn with some
offset from the start and end of the line object data. For pointed
line ends, it adds the ability to use different lengths at the start
and at the end.

This feature requires a change to the XML file format. For now, files
continue to be written in a way which allows them to be opened in
older versions of Mapper or third-party software which do not support
offsets or different-length pointed line ends. For pointed line ends,
such software will see the average length. In Mapper version 1.0,
the backwards-compatible writing feature will be removed.

Resolves #1125.