Symbol: Improve line symbol icons
This change improves two aspects. First, it changes the handling of
breaks in dashed line symbols so that the breaks are visible in most
cases, even for round or square line caps. Second, it modifies the
handling of pointed line ends and of offsets at line ends, again
contributing to the visiblity especially of dashed line symbols.
The effect is clearly visible with the line-symbol-cap-variants
test file.