HomeScreenWidgetMobile: Major rewrite
No longer show a single list of all maps in known locations,
but:
1. List most recently used maps, as on desktop.
2. If there are no such maps, show a link to the manual page.
3. List the known storage locations (unique full path).
4. Show a link to the examples folder.

When clicking on a folder, the folder contents will be displayed as
follows.
1. ".." entry which navigates back.
2. Map files in this directory.
3. Subdirectories.

This change does not allow browsing to other locations.

Geospatial vector files will no longer be listed when they are
handled by GDAL/OGR.
The "File list" headline and the redundant "Examples" button are
removed.

Closes #904. However, this change does not add custom filtering or
sorting, but just changes the way file types and folder structure
are exposed to the user.