Android: Trigger MediaScanner after saving
This is an attempt to resolve #1115, files being corrupt after
editing on Android and transferring back to the PC. In several cases,
the files had correct but incomplete data, and/or transfers after an
Android reboot were complete. This reboot effect was observed before
for README files put into the data folders, and solved by
Android::mediaScannerScanFile(). So there is a chance that using the
same pattern might eventually resolve #1115, by forcing the media
scanner to take immediate notice of the changed file. (Note that we
use QSaveFile which does actually not update the existing file but
replace it with a new file by renaming.)