GPSDisplay: Remove Positioning dependency from header
The usage of Qt Positioning is already encapsulated in the
implementation of this class. Having the layout of the class
change by presence of QT_POSITIONING_LIB required users of this
class to have the same presence of this macro, in order to avoid
memory related bugs. With this change, this transient dependency
is removed.