DrawPathTool: Fix corner point handling
Corner points are drawn by making a click-and-drag movement twice from the same
place. The first movement defines shape of the previous segment and the second
one gives direction to the next one. As there is mouse button release event at
the end of the first movement, it makes no sense to handle the case when
the click-and-drag movement returns back to its start.

Broken since 14d7daa5e4b0 ("DrawPathTool: Review undoLastPoint() et al.")
Closes GH-1176