MapPrinter: Scale painter, not resolution, for GDI
Using a very high resolution is needed for accurate vector printing
via GDI. But it must not cause the raster buffers for the background
to use a higher resolution than configured by the user.
This change needs follow-up changes in the printing code where
additional transformations must be combined with the device painter's
given transformation.