TemplateMap: Fix RenderConfig::scaling
When drawing the map, RenderConfig::scaling must be set to the number
of pixels per mm. Fixes #1186.
Document explicitly that scaling means pixel per mm.
According to the documentation, this variable is used in connection
with RenderConfig::ForceMinSize which is normally not used for
printing or export. However this change still sets the pixel per mm
from the physical (alternatively logical) resolution of the paint
device.