OcdFileImport: Handle improperly closed path segments
For objects with area-like symbols, all path segments must be properly
closed in order to produce full outlines. However, OCD data may come
with segments which lack a final point matching the first point. In
this case, we need to insert the missing point. Fixes GH-1198.

In addition, we must slightly modify the handling of the hole point
flag of properly closed segments, or the file format unit test would
stop on unexpected data changes.