AutosaveTest: Reduce flakiness
For the tests, we use an autosave interval close to one second
in order to reduce the total runtime. However, Autosave is based
on a non-precise timer, i.e. the accuracy of the timer varies
across systems. Thus test steps based on time must not only be
quick but also handle a certain variation in timing.

This change leverages QTRY...WITH_TIMEOUT to quickly discover
expected changes within a larger timeout, instead of waiting
a shorter fixed time period before checking state once.