SymbolRenderWidget: Add YesToAll, Cancel to deletion
When deleting multiple symbols, users may be asked very often to
confirm deleting the affected objects. This changes adds "Yes to all"
and "Cancel" options which will both cause the dialog to not appear
again for the current operation.

For "Cancel" to affect the whole operation, the confirmation request
is moved to the first sub-step where the symbols are collected for
later deletion.

Resolves GH-1200.
Proposed-by: @lpechacek