Compass: Fix CompassPrivate lifecycle
Random crashes appeared when exiting the app after using the
compass. CompassPrivate and its QSensor members used to be held
permanently by a static Compass instance, moving it is
destruction to after exiting the main event loop. This may be
a source for various issues, depending on Qt Sensors backend.
We can easily avoid such problems by coupling CompassPrivate
lifetime to the enabling and disabling of the compass.