MainWindow: Start an Android foreground service
To protect unsaved map changes and to keep continuity of track
recording when the screen is off or the app is in the background,
let MainWindow start an Android foreground service when the it
is told that the map is modified.
This is expected to resolve #576. However this needs to be verified
in the field, across build and device configurations.