Android: Fix BuildConfig and R access
When appending a suffix to the app's ID, the generated classes R
and BuildConfig are no longer in the org.openorienteering.mapper
namespace. For now, we can just inject additional import statements
in the build-time copy of the source tree. In the long run, this
might be solved using gradle flavours for adjusting the app ID
after compilation.