Android: Request to ignore battery optimizations
Another step to resolve GH-576 (track recording interrupted when the
screen is turned off).
This may block GH-607 (publish Mapper in Google Play Store). In this
case, the new permission should be removed from the manifest, and the
request should be called without the app ID, just guiding the user to
the relevant settings. (A toast message might give instructions to
the user.)