ActionGridBar: Refactor getButtonForAction()
This function used to return nullptr if the button couldn't be
displayed normally due to lack of screen space. However, most calls
to this functions didn't check the return value and thus might end up
dereferencing nullptr when runtime configuration (e.g. button size)
changed.
Now this function will always return the button if it was added before
(i.e. it no longer depends on runtime properties). A separate function
allows to test if a button is displayed normally or moved to the
overflow menu.