Unit tests: Link to Mapper_Common in release builds
When Mapper_DEVELOPMENT_BUILD is disabled, the additional explicit
sources are ignored, and the unit test executables are linked
directly to the Mapper_Common libary. This is a faster choice for
one-time builds like CI. It also avoids problems with collecting
test coverage information when the same source is compiled
independently for multiple targets.