Revise Android build
Allow configuration of the application ID. By default, the app ID
is 'org.openorienteering.mapper', and a '.developer' suffix is
appended when the display version differs from the plain version.
(This happens when the display version is set explicitly, or the
build is not a release configuration.) Note that the suffix is
different from what it used to be, to avoid the clash with the
signed 'dev' releases from CI.

The generated qmake configuration will now always pick up all
original sources. Source post-processing is done in the (CMake)
install step.

The 'android' folder is replaced by folders in 'src' and 'packaging'.
The launcher icons are taken directly from 'images/mapper-icon'.

Closes GH-1306.