Template: Always update state in file lookup
tryToFindTemplateFile() could be called multiple times with different
directories until successful (and so it is, in Importer::validate()).
In this situation, the final state must be Invalid iff all attempts
failed, or Unloaded iff one call succeeded to find the file. However,
the Loaded state must not be changed here because this would cause
inconsistencies with other data held by templates in this state.

Before this change, the function did not switch the state from Invalid
to Unloaded. This omission made Importer::validate() results
inconsistent dependening an template visibility settings:
- Visible templates would always be loaded immediately, based on the
  return value of the successful call to tryToFindTemplateFile().
- Hidden templates would remain in the Invalid state from the first,
  unsuccessful call to tryToFindTemplate().

This didn't cause problems on the same device because already the
first call would succeed, with the absolute path of the data. But it
most prominently failed when moving a map project to an Android device
with "all templates hidden" (sub-issue of GH-1325).