Settings: Handle touch mode besides mobile mode
Add a regular setting for a touch mode, to be actived by users on
tablet PCs. This mode shall always be active on Android, or when
mobile mode is enforced via the MAPPER_MOBILE_GUI environment
variable.

This setting uses regular members in the settings class, not the hash
based-approach of the existing settings which would add run-time
overhead for little benefit.