GeoreferencingTest: Fix angular error
When testing the calculated longitude, we must *multiply* the angular
error with a factor derived from the latitude, and we must pass the
latitude in *radians* to the cos function.
For the "LVermGeo RLP Pruem GK2" dataset, the error is slightly larger
than expected, so we need to increase the tolerance as well.