ISOM 2017-2: Add "Black for Open Land"
This color is just above yellow, for implementing black dots for
"Cultivated Land" and "Sandy ground".

This implementation is based on, but intentionally deviates at
least from the naming of, "Black for Cultivated Land" in IOF Map
Commission's "Colours ISOM 2017-2.pdf". This implementation is chosen
to allow green patches and similar on top of the named symbols.