XmlElementWriter: Write coords more efficiently
Circumvent QXmlStreamWriter when writing coordinates in compact form
to QIODevice. We can convert them directly to the UTF-8 format which
needs to go to the output device. Thus we can avoid the extra effort
for escaping special XML characters and for converting the UTF-16
QString to UTF-8 for output.
In CoordXmlTest::writeFastImplementation(), this change gives double
speed. This is significant enough to justify the unusual approach
which is tightly coupled to the actual implementation of
QXmlStreamWriter.