FileFormatTest: Cover GH-513 for text boxes
Mapper supports two variants of text objects: single anchor and box.
At the moment, the size of the box variant is saved as another
coordinate. This needs special treatment when loading (legacy) files
with out-of-bounds coordinates (issue GH-513) because the size must
not be treated like regular coordinates.
This change adds test coverage for the size of text boxes.