Symbol: Generalize `isRotatable` property
User-defied availability of object rotatability for a particular
symbol used to be an explicit PointSymbol/PointObject feature,
but needs to be applied to TextSymbol/TextObject, too (GH-1241).
As a first step, this change refactors the property originally
found in PointSymbol and moves it to class Symbol. Note that
making the setter public and saving/loading remain duties of the
symbol classes which do make use of this feature.