ISOM 2017-2: Lowest green for orchard, vineyard
The green dots (orchard) and dashes (vineyard) are used strictly with
the yellow background only. Thus the green color must be as close as
possible to the yellow color in order to give consistent covering
with other symbols such as vegetation, roads and paved area. The new
Green is also below Brown and Blue.
The changes applies to ISSkiOM 2019 as well.