cove: Find CMake package per Qt module
find_package with components is not going to actually look for the
package again when it was looked up before. In the OpenOrienteering
Mapper project, Qt5Core was already looked up from the root
CMakeLists.txt. So we need to use the same per-component package
lookup style (and Qt version) here.
Fixes a CMake message that the required package Qt was not found.