cove: Revise concurrent processing
Create a generic framework for running jobs in the background and
concurrently. Based on QtConcurrent::run, this framework adds
efficient and safe exchange of progress information across threads.

Update existing parallel image processing to use this framework:
No longer inherit from a common class, but use individual operator()
which passes the variable data to the copiable, const functor.
Add a variant of QImage which allows copying objects created for
external data without detaching on modification.

Use and maintain QThreadPool's max threads count, because this is
what QtConcurrent's thread control is based on.