DrawCircleTool: Prevent zero-sized circles
Clicking twice at the same place in the map creates a zero-sized
circle. It is reasonable to assume that this is a mistake, and thus
prevent adding the object to the map.
This change also catches the initial state when drawing is finished
before a second point is recorded.
Fixes GH-1497 (crash after drawing with circle tool).