DrawLineAndAreaTool: Allow finishDrawing() on empty path
FinishDrawing() is meant to insert the drawn (preview) object into
the map. This assumes that the object is in a valid state, esp. that
it not an empty path. Otherwise subsequent operations on the map may
crash. However, due to the various possibilities to interact with
derived tools, there may be situations where this requirement is not
satisfied. For a reliable user experience, it is necessary to handle
these situations gracefully.
Fixes GH-1497 (crash after drawing with circle tool).