LineSymbol: Ignore border color priority in equalsImpl
In contrast to 'operator==', Symbol::equals() and the equalsImpl
functions are meant to be used when comparing symbols in different
maps, e.g. during map data import in order to merge "duplicate"
symbols. For symbols in different maps, we must not compare color
priorities. This was handled incorrectly for line symbol borders.
To fix this behaviour, it is neccessary to revise the comparison
operations for LineSymbolBorder. The new implementation uses the
same duality of 'operator==' and member function 'equals' as
class Symbol.