MapColor: Improve comparison API consistency
Following the pattern from class Symbol, the public API for comparing
colors is now limited to operator==, which does also compare color
priorities, and a member function equals(), which does not look at
priorities. (There is also a static member function equal() which
accepts pointers to MapColor, but otherwise works like equals().)

The actual comparison implementation remains in a now-protected
overload of equals() which takes a bool parameter indicating whether
to compare priorities or not.