SymbolRuleSet: Revise apply() and squeezed()
Add a deleted apply(Map, Options) to prevent accidental attempts
to pass Options via {} as second argument.
Add an && qualified three-argument overload based on the new
squeeze(), and use it to implement the now-const other overload.
Always handle the RemoveUnusedSymbols flag, consistent with the
RemoveUnusedColors flag.
Base squeezed() on a new squeeze() method.