cove: Test and fix HorizontalStripes::makeStripe()
Handling color tables turned out to be not working as expected
on the stripes which were meant to share the same const data.
Based on insight into the implementation of QImage construction
and assignment, it is possible to achieve the desired behaviour
with another heap-allocated QImage sharing the same data.