cove: Remove ImageView::iw member
There is no point (and even danger) in having the image widget in a
unique_ptr because its ownership is taken by ImageView anyway when
calling QScrollArea::setWidget(). The fact that the widget() is a
non-null ImageWidget is an invariant of ImageView (at least after
construction and before destruction), so casting the result of
widget() can replace using the former iw member.
In addition, construction can be refactored to directly use an
ImageWidget object, facilitating overrides as in PolyonView.