OCD: Silence false warnings on hatched pattern area export
Mapper has a flexible area pattern definition that cannot translate to OCD in
general. However, there is one case when Mapper and OCD definitions match.
Detect the case and silence the export warning in this case. As a result, it is
possible to load an OCD file and save it again without warnings.