Georeferencing: Revise Android PROJ resource handling
The finder function only needs to copy the files to a cache directory.
PROJ can find them there if the search path points to this directory.
This fixes some use cases which need to load two or more resource
files (e.g. "+nadgrids=...") and did't work in the previous
implementation, probably due to reusing the same storage for the file
path.