Importer: Resolve template placeholders
The main import step creates TemplatePlaceholder objects when the
actual template type can not be determined immediately. In turn,
it is neccessary to do another attempt to identify the actual
template type in the validation step, when the actual template
file paths are resolved with regard to the path of the map.
Amends 92528c4, e988d53c.