GeoreferencingDialog: Fix switch to local state
Map content, including non-georeferenced templates, may be rotated
and scaled before applying the new georeferencing. However, templates
may change to non-georeferenced state when the new georeferencing is
"local". In order to have these templates rotated and scaled, too, we
must first set an intermediate georeferencing which has the same
rotation and scale as the original one, but the "local" state which
triggers the template state change.