RotateMapDialog: Always offer template checkbox
When changing map georeferencing state to local, and modifying
declination at the same time, the dialog might be called before some
templates change to non-georeferenced mode (as a result of the map
georef change). The current code fails to recognize the relevance
of template rotation in this situation. However, there is little to
gain from disabling this checkbox or from using another initial state
than checked. The checkbox would also be already enabled if there were
only closed templates which the user is not aware of.