Map: Allow findTemplateIndex on invalid templates
For simplicity, let's allow findTemplateIndex to return -1 even
in debug builds so that the calling side may decide if a missing
template is an error or not.
For example, it is okay to receive change notifications from closed
templates, but it indicates an error to have a template visibility
for a template which is not in the regular list.