ProjTransform: Construct CRS-to-CRS transforms
With modern PROJ, CRS-to-CRS transforms no longer use WGS84 as
pivotal CRS, allowing access to more precise transformations.
This change adds two new constructors. The constructor taking
two CRS specs is late-binding with modern PROJ, i.e. the actual
transformation is chosen by PROJ depending on the input coordinates
(cf. proj_create_crs_to_crs()).

When constructing transforms with implicit geographic CRS, we re-use
the CRS-to-CRS constructor. For legacy PROJ, this means to put the
geographic CRS object into the source_pj member variable, making the
static variables in ProjTransform::forward()/inverse() obsolete.
So this change not only increases uniformity in construction but
also in in transformation.

ProjTransform::crs() remains different: It is meant to be used for
cheap testing of CRS properties, not for transforming coordinates.
To prevent undefined behaviour (when used wrongly), it uses the
same PROJ object for both member variables.