Map: Remove responsibility for placeholder update
MapWidget implements the condition for displaying the placeholder
text which appears when the map doesn't contain colors, symbols, or
templates. Class Map already provides signals for changes in the
availability of these features so MapWidget needs to listen to the
signals instead of having class Map make explicit update calls.