OcdFileExport: Improve shifted rows pattern export
Mapper has a flexible area pattern definition that does not translate
to OCD in general. However, the special case of "shifted rows" can
be recognized, and subsymbols of row structures can be adjusted to
match the OCD model. In this case, no warning needs to be issued.

Closes GH-1523 (silence false warnings).

In another commit, this will be complemented by changes which
improve "aligned rows" structures.