Map: Revise template API, part 5
Update the index of the first front template when templates are
added or removed in the background of the map.
Introduce the special value '-1' to insert templates just below
the map.
Rewrite ReopenTemplateDialog::OpenTemplateList::dropEvent to make
use of this API change.

Note that the changed interactions at template deletion require
another TemplateListWidget connection to be marked as queued: When
the template table selection model emits selectionChanged, the
template table still holds the row for the template which is already
removed from Map, so table row <-> template index mappings do not
work correctly in this moment.