TemplateListWidget: Factor out TemplateTableModel
This change leverages Qt's model/view classes, separating the
template list GUI from the the data model.

To let the model properly deal with templates being added to the map
next to the map layer, the map must also explicitly signal whether a
template is meant to be added to the background or not. To avoid
a hard dependency on map.h in target slot declarations, we use
a simple bool, not a more verbose enum in map.h.