Template: Remove parameter of loadTemplateFileImpl
The configuring state is now reflected explicitly and does not need
a parameter. In addition, some code can be moved to
finishTypeSpecificTemplateConfiguration() in following commits.
For some template classes, is_georeferenced can only be true if the
configuration was loaded, so it implies that setup is not taking
place when loadTemplateFileImpl() is called.